added orders proto file

+++ order_service/internal/order/models.go
@@ -4,13 +4,13 @@ import "time"
 
 type CreateOrderDTO struct {
 	BookID string
-	UserID int
+	UserID uint64
 }
 
 type Order struct {
-	ID      int64     `db:"id"`
+	ID      uint64    `db:"id"`
 	BookID  string    `db:"book_id"`
-	UserID  int       `db:"user_id"`
+	UserID  uint64    `db:"user_id"`
 	AddedAt time.Time `db:"added_at"`
 	Status  string    `db:"status"`
 }
